Make ConnectionManager more general.
This commit changes ConnectionManager from being a static to being a
generic class that can handle connections to arbitrary jdbc URLS. As
part of this change, all Book*er classes were modified to take a
ConnectionManager in their constructor to allow for this change. Note
that, in the future, this also allows the *er classes to pre-prepare
statements rather than preparing them each time for their methods.